+++ fires3/domain/models.py
