+++ pyproject.toml
@@ -11,7 +11,8 @@ requires-python = ">=3.10"
 dependencies = [
     "boto3 (>=1.37.38,<2.0.0)",
     "firebase-admin (>=6.7.0,<7.0.0)",
-    "click (>=8.1.8,<9.0.0)"
+    "click (>=8.1.8,<9.0.0)",
+    "google-cloud-storage (>=3.1.0,<4.0.0)"
 ]
 
 [tool.poetry]
